dcos_requirements: fix centos7 install
Docker has published 19.03 in the stable repo. Since the SPEC for
`docker-ce` does not require a specific verion, when installing
`docker-ce-18.09.2` `docker-ce-cli-19.03.9` gets installed. This leads
to a `DOCKER_API_VERSION` mismatch unless the env is updated to pin the
API version. Instead we should explicitly install the matching CLI
version.